git-svn-id: svn://123.57.173.36/meijia-parent@1589 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-app/src/main/java/com/simi/action/app/dict/DictController.java
@@ -71,7 +71,7 @@ public class DictController<T> {
 
 	@RequestMapping(value = "get_ads", method = RequestMethod.GET)
     public AppResultData<Object> getAds(
-    		@RequestParam(value = "ad_type", required = false, defaultValue = "0") Short adType
+    		@RequestParam(value = "ad_type", required = false, defaultValue = "0") Long adType
     		) {
 
     	AppResultData<Object> result = null;
